moved httpClient to utils; update paths
